test(dataprotection): merge volume populator tests
